+++ src/hooks/account/useGetBatchWebsocketEvent.ts
@@ -1,0 +1,6 @@
+import { useSelector } from 'reduxStore/DappProviderContext';
+import { websocketBatchEventSelector } from 'reduxStore/selectors';
+
+export const useGetBatchWebsocketEvent = () => {
+  return useSelector(websocketBatchEventSelector);
+};