+++ src/hooks/transactions/useTransactionsTracker.ts
@@ -1,6 +1,8 @@
 import { useEffect } from 'react';
 import { getTransactionsByHashes as defaultGetTxByHash } from 'apiCalls/transactions';
 import { TransactionsTrackerType } from 'types/transactionsTracker.types';
+import { getWebsocketUrl } from 'utils/websocket/getWebsocketUrl';
+import { useGetNetworkConfig } from '../useGetNetworkConfig';
 import { useRegisterWebsocketListener } from '../websocketListener';
 import { useCheckTransactionStatus } from './useCheckTransactionStatus';
 import { useGetPollingInterval } from './useGetPollingInterval';
@@ -8,6 +10,8 @@ import { useGetPollingInterval } from './useGetPollingInterval';
 export function useTransactionsTracker(props?: TransactionsTrackerType) {
   const checkTransactionStatus = useCheckTransactionStatus();
   const pollingInterval = useGetPollingInterval();
+  const { network } = useGetNetworkConfig();
+
   const getTransactionsByHash =
     props?.getTransactionsByHash ?? defaultGetTxByHash;
 
@@ -21,11 +25,22 @@ export function useTransactionsTracker(props?: TransactionsTrackerType) {
 
   useRegisterWebsocketListener(onMessage);
 
-  useEffect(() => {
+  const setPollingInterval = async () => {
+    const websocketUrl = await getWebsocketUrl(network.apiAddress);
+
+    if (websocketUrl != null) {
+      // Do not setInterval if we already subscribe to websocket event
+      return;
+    }
+
     const interval = setInterval(onMessage, pollingInterval);
 
     return () => {
       clearInterval(interval);
     };
+  };
+
+  useEffect(() => {
+    setPollingInterval();
   }, [onMessage]);
 }